+++ src/utils/throwError.ts
@@ -1,0 +1,3 @@
+export function throwIteratorAccessError() {
+  throw new RangeError('Iterator access denied!');
+}